+++ crates/e2e/tests/e2e/eth_flow.rs
@@ -1,5 +1,5 @@
 use {
-    crate::{helpers::*, local_node::TestNodeApi},
+    crate::{local_node::TestNodeApi, setup::*},
     anyhow::bail,
     autopilot::database::onchain_order_events::ethflow_events::WRAP_ALL_SELECTOR,
     chrono::{DateTime, NaiveDateTime, Utc},
