+++ crates/e2e/tests/e2e/limit_orders.rs
@@ -1,5 +1,5 @@
 use {
-    crate::helpers::*,
+    crate::setup::*,
     ethcontract::prelude::U256,
     model::{
         order::{OrderBuilder, OrderClass, OrderKind},
