+++ crates/e2e/tests/e2e/order_cancellation.rs
@@ -1,5 +1,5 @@
 use {
-    crate::helpers::*,
+    crate::setup::*,
     ethcontract::prelude::U256,
     model::{
         app_id::AppId,
