+++ crates/e2e/tests/e2e/refunder.rs
@@ -1,8 +1,8 @@
 use {
     crate::{
         eth_flow::{EthFlowOrderOnchainStatus, ExtendedEthFlowOrder},
-        helpers::*,
         local_node::TestNodeApi,
+        setup::*,
     },
     chrono::{DateTime, NaiveDateTime, Utc},
     ethcontract::{H160, U256},
