+++ crates/e2e/tests/e2e/smart_contract_orders.rs
@@ -1,5 +1,5 @@
 use {
-    crate::helpers::*,
+    crate::setup::*,
     contracts::{GnosisSafe, GnosisSafeCompatibilityFallbackHandler, GnosisSafeProxy},
     ethcontract::{Bytes, H160, H256, U256},
     model::{
